mv task def to common use case (harvest), import to less common (fetch)
